+++ config/runtime.exs
@@ -3,7 +3,7 @@ import Config
 if config_env() == :test do
   # port = 3333
   default_database_url =
-    "postgresql://postgres:password@localhost:55555/phoenix_sync?sslmode=disable"
+    "postgresql://postgres:password@localhost:54321/phoenix_sync?sslmode=disable"
 
   database_url = System.get_env("DATABASE_URL", default_database_url)