+++ config/test.exs
@@ -17,7 +17,7 @@ db_config = [
   password: "password",
   hostname: "localhost",
   database: "phoenix_sync",
-  port: 55555
+  port: 54321
 ]
 
 # configure the support repo with random options so we can validate them in Phoenix.Sync.ConfigTest